Bump to rules_go 0.58.2, update Go in `WORKSPACE`
Updates `rules_go` to 0.58.2 per bazel-contrib/rules_go#4480, and
removes the temporary `git_override`. Updates Go to 1.25.3 in
`WORKSPACE`.

Also bumps the module version in `MODULE.bazel` to 7.1.4 in advance of
the next release.

---

The Go version update in `WORKSPACE` should've been in #1778, but I
missed including it. `rules_go` is a dev dep only required for linting,
so the impact of this oversight was minimal, but still worth correcting.

Bumped the `MODULE.bazel` version number to indicate that changes from
this point will be part of the next release.